+++ Application/Handlers/Tags/DeleteTagHandler.cs
@@ -24,20 +24,23 @@ namespace Application.Handlers.Tags
             var response = new Response<TagResponse>();
             try
             {
-                var result = await _tagRepository.GetByIdAsync(request.TagId);
-
+                var result = await _tagRepository.GetByIdAsync(request.Id);
+                if (result is null)
+                {
+                    throw new NullReferenceException("Not found tag");
+                }
                 await _tagRepository.DeleteAsync(result);
+
                 response = new Response<TagResponse>()
                 {
-                    StatusCode = HttpStatusCode.OK
+                    StatusCode = HttpStatusCode.NoContent
                 };
-
             }
-            catch (ApplicationException ex)
+            catch (NullReferenceException ex)
             {
                 response = new Response<TagResponse>(ex.Message)
                 {
-                    StatusCode = HttpStatusCode.UnprocessableEntity
+                    StatusCode = HttpStatusCode.NotFound
                 };
             }
             catch (Exception ex)
@@ -47,9 +50,7 @@ namespace Application.Handlers.Tags
                     StatusCode = HttpStatusCode.InternalServerError
                 };
             }
-
             return response;
-
         }
     }
 }
